requirements.txt
asgiref==3.2.10
cycler==0.10.0
Django==3.1
django-colorfield==0.4.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-dynamic-fields==0.3.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
